X-Forwarded-Proto: https means __Host- is ok
Previously, Redbean would refuse to allow Lua scripts to set cookies
with the prefix __Host- or __Secure- when the request was not over
HTTPS.  This meant that when Redbean is deployed behind a reverse proxy
such as nginx, Apache, or Caddy, applications requiring secure cookies
need to use TLS between the reverse proxy and Redbean also.  This is a
lot of overhead, especially when the two services are on the same
machine and communicating via loopback interface.  This change allows
__Host- and __Secure- cookies to be sent on TLS connections, or on
connections from a trusted IP address with X-Forwarded-Proto: https.